+++ src/particles.rs
@@ -42,7 +42,10 @@ pub fn particle_trail_bundle() -> (ParticleSystemBundle, Playing) {
     )
 }
 
-pub fn spawn_particle_poof(commands: &mut Commands, transform: Transform) {
+/// Spawn a particle poof at the transform's translation, adjusted to a Z of 10.0.
+pub fn spawn_particle_poof(commands: &mut Commands, transform: &Transform) {
+    let mut transform = transform.clone();
+    transform.translation.z = 10.0;
     commands.spawn((
         ParticleSystemBundle {
             transform,